Preparing CA generators for use in FastSim
