+++ FastSimulation/Tracking/python/SeedingMigration.py
@@ -12,7 +12,9 @@ def _hitSetProducerToFactoryPSet(producer):
         "PixelTripletHLTEDProducer": "PixelTripletHLTGenerator",
         "PixelTripletLargeTipEDProducer": "PixelTripletLargeTipGenerator",
         "MultiHitFromChi2EDProducer": "MultiHitGeneratorFromChi2",
-    }
+        "CAHitTripletEDProducer": "CAHitTripletGenerator",
+        "CAHitQuadrupletEDProducer": "CAHitQuadrupletGenerator",   
+        }
     ret = cms.PSet()
     _copy(producer, ret)
     ret.ComponentName = cms.string(_map[producer._TypedParameterizable__type]);
