+++ RecoPixelVertexing/PixelTriplets/src/CAHitTripletGenerator.cc
@@ -1,6 +1,6 @@
 #include <unordered_map>
 
-#include "CAHitTripletGenerator.h"
+#include "RecoPixelVertexing/PixelTriplets/interface/CAHitTripletGenerator.h"
 
 #include "RecoPixelVertexing/PixelTriplets/interface/ThirdHitPredictionFromCircle.h"
 
@@ -40,7 +40,10 @@ CAHitTripletGenerator::CAHitTripletGenerator(const edm::ParameterSet& cfg,
 		useBendingCorrection(cfg.getParameter<bool>("useBendingCorrection")),
 		caThetaCut(	cfg.getParameter<double>("CAThetaCut")),
 		caPhiCut(cfg.getParameter<double>("CAPhiCut")),
-		caHardPtCut(cfg.getParameter<double>("CAHardPtCut"))
+		caHardPtCut(cfg.getParameter<double>("CAHardPtCut")),
+		layerList(cfg.getParameter<std::vector<std::string>>("layerList")),
+		isFastSim(cfg.getParameter<bool>("isFastSim")),
+		layerPairs(cfg.getParameter<std::vector<unsigned>>("layerPairs"))
 {
 	edm::ParameterSet comparitorPSet = cfg.getParameter < edm::ParameterSet > ("SeedComparitorPSet");
 	std::string comparitorName = comparitorPSet.getParameter < std::string > ("ComponentName");
@@ -53,11 +56,19 @@ CAHitTripletGenerator::CAHitTripletGenerator(const edm::ParameterSet& cfg,
 }
 
 void CAHitTripletGenerator::fillDescriptions(edm::ParameterSetDescription& desc) {
+  edm::ParameterSetDescription empty;
+  empty.setAllowAnything();
+
   desc.add<double>("extraHitRPhitolerance", 0.06);
   desc.add<bool>("useBendingCorrection", false);
   desc.add<double>("CAThetaCut", 0.00125);
   desc.add<double>("CAPhiCut", 0.1);
   desc.add<double>("CAHardPtCut", 0);
+  desc.add<std::vector<std::string>>("layerList",{});
+  desc.add<edm::ParameterSetDescription>("BPix", empty);
+  desc.add<edm::ParameterSetDescription>("FPix", empty);
+  desc.add<bool>("isFastSim", false);
+  desc.add<std::vector<unsigned>>("layerPairs", std::vector<unsigned>{0});
 
   edm::ParameterSetDescription descMaxChi2;
   descMaxChi2.add<double>("pt1", 0.8);
