+++ RecoPixelVertexing/PixelTriplets/src/CellularAutomaton.cc
