+++ RecoPixelVertexing/PixelTriplets/src/CellularAutomaton.h
